feat/application-create-update-delete-api (#497)
